# specs/012-server-ingest-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GRAF Server Ingest API
  description: Backend ingest foundation for finalized local recording artifacts.
  version: 0.1.0
paths:
...
  /api/v1/calendar/providers:
    get:
      tags:
      - calendar
      summary: List Calendar Providers
      operationId: list_calendar_providers_api_v1_calendar_providers_get
      parameters:
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarProviderListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/calendar/sources:
    get:
      tags:
      - calendar
      summary: List Calendar Sources
      operationId: list_calendar_sources_api_v1_calendar_sources_get
      parameters:
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSourceListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - calendar
      summary: Connect Calendar Source
      operationId: connect_calendar_source_api_v1_calendar_sources_post
      parameters:
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectCalendarSourceRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSourceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/calendar/sources/{source_id}:
    get:
      tags:
      - calendar
      summary: Get Calendar Source
      operationId: get_calendar_source_api_v1_calendar_sources__source_id__get
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSourceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/calendar/sources/{source_id}/selected-calendars:
    patch:
      tags:
      - calendar
      summary: Select Calendar Source Calendars
      operationId: select_calendar_source_calendars_api_v1_calendar_sources__source_id__selected_calendars_patch
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectCalendarsRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSourceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/calendar/sources/{source_id}/sync:
    post:
      tags:
      - calendar
      summary: Sync Calendar Source
      operationId: sync_calendar_source_api_v1_calendar_sources__source_id__sync_post
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSyncResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/calendar/sources/{source_id}/disconnect:
    post:
      tags:
      - calendar
      summary: Disconnect Calendar Source Endpoint
      operationId: disconnect_calendar_source_endpoint_api_v1_calendar_sources__source_id__disconnect_post
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarDisconnectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/calendar/events/upcoming:
    get:
      tags:
      - calendar
      summary: List Upcoming Calendar Events
      operationId: list_upcoming_calendar_events_api_v1_calendar_events_upcoming_get
      parameters:
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpcomingCalendarEventsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/desktop/calendar/upcoming:
    get:
      tags:
      - calendar
      summary: List Desktop Calendar Upcoming
      operationId: list_desktop_calendar_upcoming_api_v1_desktop_calendar_upcoming_get
      parameters:
      - name: before_minutes
        in: query
        required: false
        schema:
          type: integer
          maximum: 1440
          minimum: 1
          default: 15
          title: Before Minutes
      - name: after_minutes
        in: query
        required: false
        schema:
          type: integer
          maximum: 1440
          minimum: 0
          default: 60
          title: After Minutes
      - name: X-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      - name: X-Organization-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Organization-Id
      - name: X-Workspace-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - name: X-Device-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Id
      - name: X-Client-Version
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Client-Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesktopCalendarPromptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CalendarProviderListResponse:
      properties:
        providers:
          items:
            $ref: '#/components/schemas/CalendarProviderPreset'
          type: array
          title: Providers
      type: object
      title: CalendarProviderListResponse
...
    CalendarSourceListResponse:
      properties:
        sources:
          items:
            $ref: '#/components/schemas/CalendarSourceSummary'
          type: array
          title: Sources
      type: object
      title: CalendarSourceListResponse
    CalendarSourceResponse:
      properties:
        source:
          $ref: '#/components/schemas/CalendarSourceSummary'
        calendars:
          items:
            $ref: '#/components/schemas/ExternalCalendarSummary'
          type: array
          title: Calendars
      type: object
      required:
      - source
      title: CalendarSourceResponse
...
    CalendarSyncResponse:
      properties:
        source_id:
          type: string
          format: uuid
          title: Source Id
        sync_state:
          type: string
          title: Sync State
        accepted:
          type: boolean
          title: Accepted
        event_count:
          type: integer
          title: Event Count
          default: 0
        safe_error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Safe Error Code
      type: object
      required:
      - source_id
      - sync_state
      - accepted
      title: CalendarSyncResponse
    ConnectCalendarSourceRequest:
      properties:
        provider_family:
          type: string
          title: Provider Family
        auth_mode:
          type: string
          enum:
          - app_password
          - manual_url
          title: Auth Mode
        display_label:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Display Label
        caldav_url:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Caldav Url
        username:
          anyOf:
          - type: string
            maxLength: 240
          - type: 'null'
          title: Username
        credential_input:
          anyOf:
          - type: string
          - type: 'null'
          title: Credential Input
        selected_provider_calendar_ids:
          items:
            type: string
          type: array
          title: Selected Provider Calendar Ids
      additionalProperties: false
      type: object
      required:
      - provider_family
      - auth_mode
      title: ConnectCalendarSourceRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SelectCalendarsRequest:
      properties:
        selected_provider_calendar_ids:
          items:
            type: string
          type: array
          title: Selected Provider Calendar Ids
      additionalProperties: false
      type: object
      title: SelectCalendarsRequest
...